US on JP.keylayout: Ukelele-made special keyboard layout for cjs.
This maps a Mac keyboard with a Japanese physical layout to have a
mostly US English layout with a few special cjs mods thrown in.

The biggest change is that the ESC and 1 keys are swapped because
the Mac Japanese keyboards do not have a backtick/半角 key to the
left of the '1' key. (However, Command-shift-1 will still give a '1'
as well.